backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.8
polygon-api-client==1.12.4
tweepy==4.14.0
requests==2.31.0
langchain-openai==0.0.2
langchain-chroma==0.1.2
chromadb==0.4.18
langchain-community==0.0.10